// src/types/Starship.ts
export interface IStarship {
    name: string;
    manufacturer: string;
    crew: string;
    created: string;
    url: string;
    passengers: string;
    model: string;
    starship_class: string;
    hyperdrive_rating: string;
    cost_in_credits: string;
    length: string;
    cargo_capacity: string;
}